requirements.txt

discord.py
aiohttp
Pillow
